+++ kinotic-js/workspace/packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kinotic-ai/core",
-  "version": "5.0.0-beta.0",
+  "version": "5.0.0-beta.1",
   "type": "module",
   "files": [
     "dist"
