+++ kinotic-js/workspace/packages/os-api/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kinotic-ai/os-api",
-  "version": "5.0.0-beta.0",
+  "version": "5.0.0-beta.1",
   "type": "module",
   "files": [
     "dist"
@@ -32,7 +32,7 @@
     "coverage": "bun test --coverage --pass-with-no-tests"
   },
   "peerDependencies": {
-    "@kinotic-ai/core": ">=5.0.0-beta.0"
+    "@kinotic-ai/core": ">=5.0.0-beta.1"
   },
   "dependencies": {
     "@kinotic-ai/idl": "workspace:*",
