+++ kinotic-js/workspace/packages/os-api/src/api/model/security/AuthType.ts
