+++ kinotic-js/workspace/packages/os-api/src/api/model/security/DelegatingParticipantIdentity.ts
@@ -1,6 +1,6 @@
-import { ParticipantIdentity } from '@/api/model/iam/ParticipantIdentity'
-import { ParticipantIdentityType } from '@/api/model/iam/ParticipantIdentityType'
-import { DelegateKind } from '@/api/model/iam/DelegateKind'
+import { ParticipantIdentity } from '@/api/model/security/ParticipantIdentity'
+import { ParticipantIdentityType } from '@/api/model/security/ParticipantIdentityType'
+import { DelegateKind } from '@/api/model/security/DelegateKind'
 
 /**
  * A client (a CLI install, an MCP host such as an LLM) a user has authorized to act on their
