+++ kinotic-js/workspace/packages/os-api/src/api/model/security/IApplicationParticipant.ts
