+++ kinotic-js/workspace/packages/os-api/src/index.ts
@@ -43,17 +43,17 @@ export * from '@/api/model/workload/VmNodeStatus'
 
 export * from '@/api/model/log/LogQuery'
 
-export * from '@/api/model/iam/AuthType'
-export * from '@/api/model/iam/ParticipantIdentity'
-export * from '@/api/model/iam/UserParticipantIdentity'
-export * from '@/api/model/iam/DelegatingParticipantIdentity'
-export * from '@/api/model/iam/ParticipantIdentityType'
-export * from '@/api/model/iam/DelegateKind'
-export * from '@/api/model/iam/DelegateSession'
-export * from '@/api/model/iam/SignUpRequest'
-export * from '@/api/model/iam/SignUpCompleteRequest'
-export * from '@/api/model/iam/CompleteOrgRequest'
-export * from '@/api/model/iam/PendingInviteSummary'
+export * from '@/api/model/security/AuthType'
+export * from '@/api/model/security/ParticipantIdentity'
+export * from '@/api/model/security/UserParticipantIdentity'
+export * from '@/api/model/security/DelegatingParticipantIdentity'
+export * from '@/api/model/security/ParticipantIdentityType'
+export * from '@/api/model/security/DelegateKind'
+export * from '@/api/model/security/DelegateSession'
+export * from '@/api/model/security/SignUpRequest'
+export * from '@/api/model/security/SignUpCompleteRequest'
+export * from '@/api/model/security/CompleteOrgRequest'
+export * from '@/api/model/security/PendingInviteSummary'
 export * from '@/api/model/InviteEmailTemplate'
 
 export * from '@/api/model/github/GitHubAppInstallation'
@@ -64,12 +64,12 @@ export * from '@/api/model/github/GitHubInstallCompletion'
 export * from '@/api/model/KinoticProjectConfig'
 
 // Security
-export * from '@/api/security/ParticipantType'
-export * from '@/api/security/ISystemParticipant'
-export * from '@/api/security/IOrganizationParticipant'
-export * from '@/api/security/IApplicationParticipant'
-export * from '@/api/security/ParticipantGuards'
-export * from '@/api/security/KinoticOsCredentialsAuthProvider'
+export * from '@/api/model/security/ParticipantType'
+export * from '@/api/model/security/ISystemParticipant'
+export * from '@/api/model/security/IOrganizationParticipant'
+export * from '@/api/model/security/IApplicationParticipant'
+export * from '@/api/model/security/ParticipantGuards'
+export * from '@/api/model/security/KinoticOsCredentialsAuthProvider'
 
 // Services
 export * from '@/api/services/IApplicationService'
