+++ website/content/02.platform/11.reference/01.cri-format.md
@@ -140,7 +140,7 @@ cri.raw()           // 'srv://tenant-123@app.acme-org.orders-app~OrderService/pl
 
 | CRI | Description |
 |-----|-------------|
-| `srv://os-api~org.kinotic.os.api.services.iam.MemberService` | A platform service in the `os-api` zone |
+| `srv://os-api~org.kinotic.os.api.services.security.MemberService` | A platform service in the `os-api` zone |
 | `srv://app-api~org.kinotic.persistence.api.services.JsonEntitiesRepository/save` | A specific method on a platform service |
 | `srv://app.acme-org.orders-app~OrderService/create#1.0.0` | A versioned method on an application's own service |
 | `srv://app.acme-org.orders-app.billing~InvoiceService` | A service in an application-declared sub-zone |
